+++ packages/kaos/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add `env` parameter to `exec()` method for passing environment variables to subprocesses
+
 ## 0.6.0 (2026-01-09)
 
 - Add optional `n` parameter to `readbytes` to read only the first n bytes